Tests/LibWeb/Text/expected/form-elements-keeps-form-alive.txt

form alive while collection is held: true
collection.length: 0
